Reorganize constraints and transforms modules
