+++ pyro/distributions/torch_distribution.py
@@ -1,9 +1,9 @@
 import warnings
 
 import torch
-from torch.distributions import biject_to, constraints, transform_to
 from torch.distributions.kl import kl_divergence, register_kl
 
+import pyro.distributions.constraints as constraints
 import pyro.distributions.torch
 from pyro.distributions.distribution import Distribution
 from pyro.distributions.util import broadcast_shape, scale_and_mask
@@ -196,33 +196,6 @@ class TorchDistribution(torch.distributions.Distribution, TorchDistributionMixin
     ``.has_enumerate_support = True``.
     """
     pass
-
-
-# TODO move this upstream to torch.distributions
-class IndependentConstraint(constraints.Constraint):
-    """
-    Wraps a constraint by aggregating over ``reinterpreted_batch_ndims``-many
-    dims in :meth:`check`, so that an event is valid only if all its
-    independent entries are valid.
-
-    :param torch.distributions.constraints.Constraint base_constraint: A base
-        constraint whose entries are incidentally indepenent.
-    :param int reinterpreted_batch_ndims: The number of extra event dimensions that will
-        be considered dependent.
-    """
-    def __init__(self, base_constraint, reinterpreted_batch_ndims):
-        self.base_constraint = base_constraint
-        self.reinterpreted_batch_ndims = reinterpreted_batch_ndims
-
-    def check(self, value):
-        result = self.base_constraint.check(value)
-        result = result.reshape(result.shape[:result.dim() - self.reinterpreted_batch_ndims] + (-1,))
-        result = result.min(-1)[0]
-        return result
-
-
-biject_to.register(IndependentConstraint, lambda c: biject_to(c.base_constraint))
-transform_to.register(IndependentConstraint, lambda c: transform_to(c.base_constraint))
 
 
 class MaskedDistribution(TorchDistribution):
